Put qoiz encoding tests behind build flags as well
